fix: update serverless swagger documentation types
