--- src/types/product.ts
@@ -1,16 +1,0 @@
-export type Product = {
-  id: string;
-  title: string;
-  description: string;
-  count: number;
-  price: number;
-  image_url: string;
-};
-
-export type ProductPostBody = {
-  title: string;
-  description: string;
-  count: number;
-  price: number;
-  image_url: string;
-};
