--- src/types/stock.ts
@@ -1,4 +1,0 @@
-export type Stock = {
-  product_id: string;
-  count: number;
-};
